# specs/001-build-volunteersync-an/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VolunteerSync API
  description: |
    RESTful API for VolunteerSync volunteer management platform.
    
    **Authentication**: JWT Bearer token required for protected endpoints.
    **Rate Limiting**: 100 requests per minute per user, 5 login attempts per 15 minutes per IP.
    **Base URL**: `http://localhost:8080/api/v1` (development)
    
    ## API Modules:
    - **Authentication**: User registration, login, password reset
    - **Users**: User profile management
    - **Organizations**: Organization CRUD, team management
    - **Volunteers**: Volunteer profile management
    - **Opportunities**: Event/opportunity CRUD, search
    - **Registrations**: Volunteer event registration
    - **Hours**: Hours tracking and verification
    - **Communications**: Messages and notifications
    - **Analytics**: Reporting and dashboards
    - **Documents**: Document management
    - **Teams**: Team/group management
    - **Admin**: Platform administration
  version: 1.0.0
  contact:
    name: VolunteerSync API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://staging.volunteersync.com/api/v1
    description: Staging server
  - url: https://api.volunteersync.com/api/v1
    description: Production server
...
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Retrieve authenticated user's profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    patch:
      tags:
        - Users
      summary: Update current user profile
      description: Update authenticated user's profile information
      operationId: updateCurrentUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                  format: email
                  description: Requires reverification (FR-007)
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/delete:
    delete:
      tags:
        - Users
      summary: Delete user account
      description: Request account deletion with data retention compliance (FR-009, FR-107)
      operationId: deleteUserAccount
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account deletion requested
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # ORGANIZATIONS MODULE
  # ============================================================================
  /organizations:
    get:
      tags:
        - Organizations
      summary: List organizations
      description: Get paginated list of organizations with optional filters
      operationId: listOrganizations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          schema:
            type: string
          description: Search by organization name
        - name: cause
          in: query
          schema:
            type: string
          description: Filter by cause category slug
        - name: city
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Organizations list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create new organization profile (FR-010)
      operationId: createOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{org_id}:
    get:
      tags:
        - Organizations
      summary: Get organization by ID
      description: Retrieve organization profile (public endpoint, FR-011)
      operationId: getOrganization
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organization retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Organizations
      summary: Update organization
      description: Update organization profile (admin only)
      operationId: updateOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdate'
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

# Additional paths continue...
# Due to length, showing structure for key endpoints
# Full OpenAPI spec would include all 111 functional requirements
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15 minutes)
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Items per page
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
          nullable: true
        account_status:
          type: string
          enum: [active, inactive, suspended]
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        mission_statement:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        address:
          $ref: '#/components/schemas/Address'
        logo_url:
          type: string
          nullable: true
        banner_url:
          type: string
          nullable: true
        verification_status:
          type: string
          enum: [verified, unverified]
        total_volunteers:
          type: integer
        total_hours:
          type: number
          format: float
        avg_rating:
          type: number
          format: float
          nullable: true
        causes:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    OrganizationCreate:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        mission_statement:
          type: string
        description:
          type: string
        website:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        logo_url:
          type: string
        banner_url:
          type: string
        cause_ids:
          type: array
          items:
            type: string
            format: uuid
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
        mission_statement:
          type: string
        description:
          type: string
        website:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        logo_url:
          type: string
        banner_url:
          type: string
        cause_ids:
          type: array
          items:
            type: string
            format: uuid
    Address:
      type: object
      properties:
        address_line_1:
          type: string
        address_line_2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
          default: "United States"
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
        total_items:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: invalid_request
        message:
          type: string
          example: Validation error
        details:
          type: object
          additionalProperties: true
          nullable: true